elements: split into <patchwork-view> + <patchwork-component>, ship JSX types
- Rename old patchwork-view-legacy.ts → patchwork-view.ts: the
  doc-url/tool-id element is now just <patchwork-view> (no -legacy suffix).
- Extract the new component-rendering element into a separate
  patchwork-component.ts as <patchwork-component>, stripped of the legacy
  passthrough plumbing it previously needed to fall through to the old
  view. It only handles `component`/`url`.
- Bootloader and both sites updated to use the new names directly.

Ship JSX intrinsic types from the package itself:

- Move declarations from src/elements.d.ts into src/index.ts using
  `import type {}` side effects, so tsc emits the `declare module` /
  `declare global` blocks into dist/index.d.ts.
- Add react/@types/react/solid-js as optional peerDependencies (and dev
  deps for local typecheck). Consumers that import anything from
  @inkandswitch/patchwork-elements now get typed JSX intrinsics for
  both elements automatically; no more per-package patchwork-view.d.ts.

Tarball housekeeping:

- Move workspace tarballs from dist-tarballs/ to dist/ (already covered
  by the existing .gitignore `dist/` rule, removing the need for a
  separate dist-tarballs/ entry).
