+++ core/elements/package.json
@@ -11,7 +11,15 @@
     "@automerge/automerge-repo": "*",
     "@inkandswitch/patchwork-filesystem": "workspace:^",
     "@inkandswitch/patchwork-plugins": "workspace:^",
-    "@inkandswitch/patchwork-providers": "workspace:^"
+    "@inkandswitch/patchwork-providers": "workspace:^",
+    "@types/react": "*",
+    "react": "*",
+    "solid-js": "*"
+  },
+  "peerDependenciesMeta": {
+    "@types/react": { "optional": true },
+    "react": { "optional": true },
+    "solid-js": { "optional": true }
   },
   "dependencies": {
     "debug": "^4.4.3"
@@ -28,6 +36,9 @@
     "@inkandswitch/patchwork-plugins": "workspace:^",
     "@inkandswitch/patchwork-providers": "workspace:^",
     "@types/debug": "^4.1.12",
-    "@types/node": "^20.19.9"
+    "@types/node": "^20.19.9",
+    "@types/react": "catalog:",
+    "react": "^18.3.1",
+    "solid-js": "^1.9.10"
   }
 }
