+++ sites/tiny-patchwork/index.html
@@ -21,7 +21,7 @@
 <link rel="dns-prefetch" href="https://sync3.automerge.org" />
 <link rel="preload" href="/automerge.wasm" as="fetch" crossorigin />
 <link rel="preload" href="/subduction.wasm" as="fetch" crossorigin />
-<patchwork-view-legacy id="root"></patchwork-view-legacy>
+<patchwork-view id="root"></patchwork-view>
 <script type="module" src="src/main.ts"></script>
 
 </html>
